Add slash to SafetyModuleHandler

+++ src/lib/structs/Slash.sol
@@ -3,5 +3,5 @@ pragma solidity 0.8.22;
 
 struct Slash {
   uint16 reservePoolId;
-  uint128 amount;
+  uint256 amount;
 }
